@@ -1,18 +1,19 @@
 {
   "name": "architect-typescript-example",
   "scripts": {
+    "lint": "eslint ./src --fix",
     "start": "arc sandbox"
   },
   "dependencies": {
-    "@architect/plugin-typescript": "^0.1.0",
     "whatev": "^1.0.0"
   },
   "devDependencies": {
-    "@architect/architect": "^10.0.3",
+    "@architect/architect": "^10.2.2",
     "@architect/eslint-config": "^2.0.1",
+    "@architect/plugin-typescript": "^0.1.0",
     "@types/aws-lambda": "^8.10.93",
-    "@types/node": "^17.0.21",
-    "eslint": "^8.10.0"
+    "@types/node": "^17.0.23",
+    "eslint": "^8.13.0"
   },
   "eslintConfig": {
     "extends": "@architect/eslint-config"
