@@ -1,3 +1,3 @@
-module.exports = function (): string {
-  return `this bit of text was exported from custom TS shared`
+export default function (): string {
+  return `this bit of text was exported from custom TS shared`;
 }
